+++ filen-cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- /
+
+## 0.2.7 - 2026-06-19
+
 ### Added
 
 - `view-html-docs` command to view the in-app docs locally in a browser rendered as HTML